handles events correctly and ties them to simple
`once` sets an event handler for removal, after handling the event
once. this caused `show`'s handler to be called once, in which case,
if handling the showing of another file dialog (e.g., save), it
wouldn't be called again (e.g., to handle the showing of the "open
file" dialog, if shown afterwards).

on the other hand, we want `afterChoose`'s handler to be removed once
handled (in order not to affect other file dialogs), and added again
when the "open file" dialog is shown.

this fixes the problem by using `on` to listen for the `show` event,
and `once` to listen for the `afterChoose` event.

this also ties the added events to simple (to be removed when simple
is unloaded).