ensures div after "Preferences" shows when should
When toggling less-comfy on, some menu items are hidden which triggers
some dividers to automatically hide. When toggling more-comfy on again
the dividers that were automatically hidden are not automatically
shown again.

This fixes the problem by caching the dividers that get hidden
automatically and showing them in more-comfy only.